Split out the unscented correction function/structure into separate file and update sigma point generator.
